home work 6

+++ database/migrations/2022_10_10_133913_create_news_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('news', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('category_id');
+            $table->integer('category_id');
             $table->string('title', 150);
             $table->text('text');
             $table->timestamps();
